Produce a single interface instead of nested interfaces.
It turns out nested interfaces are not supported (such nesting
of namespacing should be handled by the package definition only)
so a fairly major rewrite was needed.  This also necessitated the
introduction of some new concepts for full flexibility, namely to
strip-off part of the prefix so it is not always put into the interface
name.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>